Relocated functions. Added movement logic. New file structure.
