+++ ChessGame/__main__.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# from src.ChessObj import *
+
+def main():
+    """ Main program """
+    # Code goes over here.
+    return 0
+
+
+if __name__ == "__main__":
+    main()
