+++ ChessGame/src/Sprites/cursed_sheet.png
